Modularized the controller, each evaluation method now has its own controller.py file
