+++ tethysapp/community_streamflow_evaluation_system/app.py
@@ -11,8 +11,14 @@ class CSES(TethysAppBase):
     index = 'home'
     icon = f'{package}/images/CSESoverviewImage.JPG' 
     root_url = 'community-streamflow-evaluation-system' 
-
     color = '#003087'
+    description = (
+        '"A tethys app that lets the user to visualize and query modeled streamflow'
+    )
+    tags = '"Hydrology", "WMO", "BYU"'
+    enable_feedback = False
+    feedback_emails = []
+    controller_modules = ["controllers", "State_Controller", "Reach_Controller", "HUC_Controller"]
  
     def custom_settings(self):
         """
